modules/core/petra/petra-lang/src/main/java/com/liferay/petra/lang/ClassResolverUtil.java
/**
 * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

package com.liferay.petra.lang;

import java.util.HashMap;
import java.util.Map;

/**
 * @author Shuyang Zhou
 */
public class ClassResolverUtil {

	public static Class<?> resolve(String className, ClassLoader classLoader)
		throws ClassNotFoundException {

		try {
			return Class.forName(className, false, classLoader);
		}
		catch (ClassNotFoundException classNotFoundException) {
			Class<?> clazz = _primitiveClasses.get(className);

			if (clazz != null) {
				return clazz;
			}

			throw classNotFoundException;
		}
	}

	private static final Map<String, Class<?>> _primitiveClasses =
		new HashMap<String, Class<?>>(9, 1.0F) {
			{
				put("boolean", boolean.class);
				put("byte", byte.class);
				put("char", char.class);
				put("double", double.class);
				put("float", float.class);
				put("int", int.class);
				put("long", long.class);
				put("short", short.class);
				put("void", void.class);
			}
		};

}